Refactor code for the demo site
